requirements.txt
# Core AI/ML Dependencies
langchain>=0.1.14
langchain-google-genai>=0.0.12
langchain-openai>=0.1.4
langgraph>=0.0.25

# LLM & Vector Database
google-generativeai>=0.3.0
openai>=1.3.0
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2

# Market Data APIs
yfinance>=0.2.32
alpha-vantage>=3.0.0
requests>=2.31.0

# Web Framework & UI
streamlit>=1.28.1
plotly>=5.17.0
pandas>=2.1.0
numpy>=1.26.0

# Data Processing & Storage
json5>=0.9.14
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Testing & Quality
pytest>=7.4.3
pytest-cov>=4.1.0
pytest-mock>=3.12.0
black>=23.12.0
flake8>=6.1.0

# Logging & Monitoring
python-json-logger>=2.0.7

# Utilities
tenacity>=8.2.3
pydantic>=2.5.0
cachetools>=5.3.2